Split visualizer helpers into modules
