+++ index.html
@@ -47,6 +47,6 @@
     </aside>
     <section id="visualizer" class="flex-1 relative overflow-hidden"></section>
   </main>
-  <script src="app.js"></script>
+  <script type="module" src="main.js"></script>
 </body>
 </html>
