+++ main.js
@@ -1,0 +1,3 @@
+import { initVisualizer } from './js/visualizer.js';
+
+document.addEventListener('DOMContentLoaded', initVisualizer);
